refactor(cli): break parent path descriptor cycle
